+++ lib_install.sh
@@ -140,7 +140,14 @@ pip3 install shodan
 pip3 install datetime
 pip3 install pytz
 pip3 install colorama
-echo '[*]Initializing mousejack...'
+echo '[*]Installing and Initializing PhoneInfoga...'
+git clone https://github.com/sundowndev/PhoneInfoga
+cd PhoneInfoga
+pip3 install future
+python3 -m pip install -r requirements.txt
+cp config.example.py config.py
+cd ../
+echo '[*]Installing and Initializing mousejack...'
 git clone https://github.com/BastilleResearch/mousejack.git
 cd mousejack
 git submodule init